Fixed some code smells detected by SonarCloud

+++ src/LibSlg/Values/Klass.hpp
@@ -14,7 +14,7 @@ public:
 
 	bool knowsIdentifier(const std::string& identifier) const;
 	const std::shared_ptr<Context>& getInstanceContext() const { return m_instanceContext; }
-	unsigned int getArity() const { return m_declarationStmt.getConstructorParameters().size(); }
+	unsigned long getArity() const { return m_declarationStmt.getConstructorParameters().size(); }
 	void instantiate(std::vector<Value::Ptr> constructorArguments);
 
 	bool isImplicitlyTrue() const override { return true; }
